pyproject.toml
[project]
name = "opencontext"
version = "1.0.0"
description = "Extensible MCP framework template for civic data platforms"
authors = [
    {name = "Srihari Raman", email = "srihari.raman@boston.gov"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0",
    "tenacity>=8.0.0",
    "python-json-logger>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/cityofboston/opencontext"
Documentation = "https://github.com/cityofboston/opencontext/blob/main/README.md"
Repository = "https://github.com/cityofboston/opencontext"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

